Create Default Cognostics Data Frame

Create default cognostics data frame to be used to convert columns in dilution_summary to class cognostics

create_default_cog_df()

Value

A dataframe of default cognostics information

Details

Internal function to create a dataframe of default cognostics information to be used by the trelliscopejs::cog.

Examples

cog_df <- create_default_cog_df()
cog_df

#>             col_name_vec                                    desc_vec type_vec
#> 1        Transition_Name                             Transition_Name   factor
#> 2         Dilution_Batch                              Dilution_Batch   factor
#> 3  Transition_Name_Class                      Classes of Transitions   factor
#> 4              wf1_group                       Group from workflow 1   factor
#> 5              wf2_group                       Group from workflow 2   factor
#> 6                 r_corr                Pearson Correlation R values  numeric
#> 7             pra_linear Linear Regression Percent Residual Accuracy  numeric
#> 8           mandel_p_val                    P values for Mandel test  numeric
#> 9              r2_linear                 Linear Regression R^2 Value  numeric
#> 10         r2_adj_linear        Linear Regression Adjusted R^2 Value  numeric
#> 11          mandel_stats             Test statistics for Mandel Test  numeric