Validate Trellis Table

Validate trellis table to see if it can be used to create trelliscopejs report

validate_trellis_table(
  trellis_table,
  grouping_variable = c("Transition_Name", "Dilution_Batch"),
  panel_variable = "panel"
)

Arguments

trellis_table
The trellis table generated by function add_plotly_panel()

grouping_variable
A character vector of column names in trellis_table to indicate how each dilution curve should be grouped by. It is also going to be used as a conditional cognostics in the trelliscopejs report. The function will check if these columns are indeed conditional cognostics Default: c("Transition_Name", "Dilution_Batch")

panel_variable
A column name in trellis_table to indicate the list of plots to be used in the trelliscopejs report. This column name must be fixed to the name panel and it must be of class trelliscope_panels Default: 'panel'

Details

A valid trellis table must have a column named panel with a list of plots and class trelliscope_panels, conditional cognostics columns as indicated in grouping_variable and the rest of the columns as common cognostics

Examples

sample_name <- c("Sample_010a", "Sample_020a", "Sample_040a",
                 "Sample_060a", "Sample_080a", "Sample_100a",
                 "Sample_010b", "Sample_020b", "Sample_040b",
                 "Sample_060b", "Sample_080b", "Sample_100b")
lipid1_area <- c(22561, 31178, 39981, 48390, 52171, 53410,
                 32561, 41178, 49981, 58390, 62171, 63410)
lipid2_area <- c(2299075, 4136350, 7020062, 8922063, 9288742, 11365710,
                 2300075, 4137350, 7021062, 8923063, 9289742, 11366710)
lipid_data <- tibble::tibble(Sample_Name = sample_name,
                             Lipid1 = lipid1_area,
                             Lipid2 = lipid2_area)
dilution_percent <- c(10, 20, 40, 60, 80, 100,
                      10, 20, 40, 60, 80, 100)
dilution_batch <- c("B1", "B1", "B1", "B1", "B1", "B1",
                    "B2", "B2", "B2", "B2", "B2", "B2")
dilution_annot <- tibble::tibble(Sample_Name = sample_name,
                                 Dilution_Batch = dilution_batch,
                                 Dilution_Percent = dilution_percent)
# Create dilution table
dilution_table <- create_dilution_table(dilution_annot, lipid_data,
                                        common_column = "Sample_Name",
                                        signal_var = "Area",
                                        column_group = "Transition_Name"
)
# Create a trellis table without dilution summary
trellis_table_auto <- add_plotly_panel(dilution_table) %>%
   convert_to_cog(cog_df = NULL,
                  grouping_variable = c("Transition_Name",
                                        "Dilution_Batch"),
                  panel_variable = "panel",
                  col_name_vec = "col_name_vec",
                  desc_vec = "desc_vec",
                  type_vec = "type_vec")

# Check if trellis_table_auto is valid
validate_trellis_table(trellis_table_auto)